MB-22008: Introduce CV Jenkinsfile
Allows the use of Multibranch Pipeline Jobs in Jenkins to centralise
configuration of the various jobs. Many of the existing jobs simply call
single-project-gerrit with different env variables to configure it.

This commit introduces a Jenkinsfile reimplementation of single-project-gerrit
and single-project-clang-static-analyzer, and a set of config files specifying
the env variables for a given job

Change-Id: Icf937a3ab8e013486e0f9b261426df42e57af901
Reviewed-on: http://review.couchbase.org/100355
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>