Use currentBuild.startTimeInMillis if real start time not set
Jenkins provides currentBuild.startTimeInMillis but calculating the job
duration using this time includes the time spent waiting for a worker
to be available. As this varies wildly, and can be unsettling to see an
overly long job time, the start of the actual build is tracked
manually.

If an error occurs, this value might not get set. In this case, falling
back on currentBuild.startTimeInMillis allows a report to be made to
Gerrit verify-status pane and avoids unclear exceptions being logged.

Change-Id: Ia3e9c30afca2c55016ffc66efee020dee737fb3b
Reviewed-on: http://review.couchbase.org/116862
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: James Harrison <james.harrison@couchbase.com>