CBD-3263: Remove per-job PARALLELISM overrides, adjust defaults
Remove the existing overloads for different job types which assumed a
homogeneous set of machines for each job type (which is no longer the
case now some CV machines are 16 core and others 4).

Change the default PARALLELISM / TEST_PARALLELISM if not otherwise set
to 6 and (PARALLELISM / 2) to behave reasonably in the smaller CV machines.

Note: this is a merge of the following 3 commits from master branch:

* 37e8f96 2020-03-02 | CBD-3263: Set default parallelism to fixed value (6) [Dave Rigby]
* 0a6a0a2 2020-03-02 | CBD-3263: Convert PARALLELISM to int before performing arithmetic [Dave Rigby]
* f1085a4 2020-03-02 | CBD-3263: Set default parallelism based on #CPUs [Dave Rigby]

Change-Id: Ifaba534273a01f904ff4334547ab1db48281b81a
Reviewed-on: http://review.couchbase.org/123157
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>