CBD-3263: Set default parallelism to fixed value (6)
It turns out that any expressions in Jenkinsfile are evaulated only
once, on the main Jenkins master server, not every time the job runs.
As such, it doesn't appear possible to dynamically set the PARALLELISM
env var based on the CPU count the job is running on.

As such, revert to a fixed default value for PARALLELISM (and
TEST_PARALLELISM) if not otherwise defined.

Instead we will set environment variables in the individual Jenkins
node config.

Change-Id: I4fd7088145d2fb220cda0ddef4f790bb8fc9f7f0
Reviewed-on: http://review.couchbase.org/123067
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>