CBD-3263: Facilitate using different build generators (e.g. Ninja)
Restructure the Jenkinsfile to split the configuration and build,
avoiding the use of the wrapper Makefile and driving cmake and the
build directly.

This allows an alternative generator (e.g. Ninja) to be specified via
the CMAKE_ARGS env var (CMAKE_ARGS="-G Ninja") while still letting the
existing builds use the default generator.

Change-Id: If54c94ee61b1b2a232896b998b03040cfff87208
Reviewed-on: http://review.couchbase.org/123141
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>