CBD-3263: Use Ninja as default build generator
(Backport of a30b8a3, aa3abca and 9e17346 to mad-hatter)

Restructure the Jenkinsfile to split the configuration and build,
avoiding the use of the wrapper Makefile and driving cmake and the
build directly. Remove any hard-coded assumptions that `make` is the
build generator (don't run tests via make, and don't test for
Makefile).

Default to using Ninja as the CMake generator; can be overridden via
the CMAKE_GENERATOR env var.

Change-Id: I9ceb49c7ade390715a1458d2e7d12b23a933c6ca
Reviewed-on: http://review.couchbase.org/c/server-cv/+/125131
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>