Add support for project-common config files
This allows a particular project (kv_engine, tlm, ...) to specify a
'common' config file to be used for all jobs assosicated with that
project.

The motivating use-case for this is to add support for 'tlm' CV jobs,
which should not have COUCHBASE_KV_COMMIT_VALIDATION set anywhere; so
we want to make that COUCHBASE_KV_COMMIT_VALIDATION only enabled for
non-tlm jobs.

Change-Id: I3ef2d624693e2e10e87d865dd055061637dd80b3
Reviewed-on: http://review.couchbase.org/c/server-cv/+/161388
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>