tlm/common: Don't override CB_DOWNLOAD_DEPS_REPO for jobs on EC2
The tlm CV job runs with an overide for CB_DOWNLOAD_DEPS_REPO to allow
one to test proposed cbdeps binaries which have not yet been
published. However the overridden URL is behind the VPN and hence is
not accessible to machines running outside the VPN - such as aarch64
Linux nodes which are EC2 instances.

Don't override the CB_DOWNLOAD_DEPS_REPO for these job types. This
does mean that such jobs will fail unless the proposed cbdep has been
published - but that compromise is the most straightfoward way to get
this working.

Change-Id: I8065ff54276f3d6bbc8c202fe2485b32cb20a4c9
Reviewed-on: https://review.couchbase.org/c/server-cv/+/183895
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Blair Watt <blair.watt@couchbase.com>