Fix for the annoying 'jdk.exploded' CV failures
The cleanWs() step was only deleting files, not the directory hierarchy.
This messes up the CBDEP_INSTALL() CMake macro, which uses the existence
of the target directory (jdk.exploded) as a short-circuit to avoid the
slow 'cbdep install' step.

Solution: delete the build/ and install/ directories outright, being
sure to specify deleteDirs: true so they actually get removed.

Change-Id: Ia340d8a0aff77ac8236c30e495bb5da0c10f4f5e
Reviewed-on: https://review.couchbase.org/c/server-cv/+/186075
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Blair Watt <blair.watt@couchbase.com>