CBD-6071: Remove verify-status actions
The verify-status plugin for Jenkins has not been updated to work with
new releases of Gerrit. We will no longer report back to gerrit this
way, but via comments (and checks-jenkins/Gerrit Checks once an
outstanding authentication issue has been ironed out)

This change also disables Gerrit Triger's silentStartMode in the
interim to avoid confusion, so that a comment will be posted on the
change detailing that the CV job has begun.

Change-Id: I67e9063f086d00d7d9c6da6a1929786023a26555
Reviewed-on: https://review.couchbase.org/c/server-cv/+/218078
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Rohan Suri <rohan.suri@couchbase.com>